+++ src/melcloudatw.js
@@ -14,7 +14,7 @@ class MelCloudAtw extends EventEmitter {
         const accountInfoFile = config.accountInfoFile;
         const deviceInfoFile = config.deviceInfoFile;
         const debugLog = config.debugLog;
-        const refreshInterval= config.refreshInterval;
+        const refreshInterval = config.refreshInterval;
 
         //set default values
         this.deviceData = {};
@@ -428,7 +428,7 @@ class MelCloudAtw extends EventEmitter {
                 const debug2 = debugLog ? this.emit('debug', `Device State: ${JSON.stringify(deviceState, null, 2)}`) : false;
                 this.emit('deviceState', deviceData, deviceState, useFahrenheit);
             } catch (error) {
-                this.emit('error', `Check device error: ${error}.`);
+                this.emit('error', `Check device error: ${error.message ?? error}.`);
             };
         }).on('state', () => { });
 
@@ -441,7 +441,7 @@ class MelCloudAtw extends EventEmitter {
             const data = savedData.length > 0 ? JSON.parse(savedData) : false;
             return data;
         } catch (error) {
-            throw new Error(`Read data from path: ${path}, error: ${error}`);
+            throw new Error(`Read data from path: ${path}, error: ${error.message ?? error}`);
         }
     }
 
@@ -468,7 +468,7 @@ class MelCloudAtw extends EventEmitter {
             this.emit('deviceState', this.deviceData, deviceState, this.useFahrenheit);
             return true;
         } catch (error) {
-            throw new Error(error);
+            throw new Error(error.message ?? error);
         };
     };
 };